+++ packages/fcl/src/exec/utils/prep-template-opts.js
@@ -1,9 +1,9 @@
-import * as sdk from "@onflow/sdk"
 import {retrieve} from "../../document/document.js"
 import {normalizeInteractionTemplate} from "../../normalizers/interaction-template/interaction-template"
 import {deriveCadenceByNetwork} from "../../interaction-template-utils/derive-cadence-by-network.js"
 import {deriveDependencies} from "./derive-dependencies"
 import {isString} from "./is"
+import {getChainId} from "../../utils"
 
 export async function prepTemplateOpts(opts) {
   if (isString(opts?.template)) {
@@ -20,7 +20,7 @@ export async function prepTemplateOpts(opts) {
     opts.cadence ||
     deriveCadenceByNetwork({
       template: opts.template,
-      network: await sdk.config().get("flow.network"),
+      network: await getChainId(),
     })
 
   opts.cadence = cadence