+++ packages/fcl/src/fcl.js
@@ -37,6 +37,8 @@ export {AppUtils}
 import * as InteractionTemplateUtils from "./interaction-template-utils"
 export {InteractionTemplateUtils}
 
+export {getChainId} from "./utils"
+
 export {TestUtils} from "@onflow/sdk"
 export {config} from "@onflow/config"
 export {send} from "@onflow/sdk"
@@ -66,6 +68,7 @@ export {getBlockHeader} from "@onflow/sdk"
 export {getCollection} from "@onflow/sdk"
 export {getTransactionStatus} from "@onflow/sdk"
 export {getTransaction} from "@onflow/sdk"
+export {getNetworkParameters} from "@onflow/sdk"
 export {authorizations, authorization} from "@onflow/sdk"
 export {args, arg} from "@onflow/sdk"
 export {proposer} from "@onflow/sdk"