+++ packages/fcl/src/utils/index.js
@@ -1,3 +1,5 @@
+export {getChainId} from "./getChainId"
+
 export function isAndroid() {
   return (
     typeof navigator !== "undefined" && /android/i.test(navigator.userAgent)